Refactor Privateer validation integration for better performance
This commit removes the linkage to `privateer-wrapper.cc` from `CMakeLists.txt`, and adds a large number of changes to `moorhen-wrappers.cc`. It focuses on adding Privateer's functions to moorhen-wrappers, resulting in a more simplified and improved validation without getting string coordinates from JavaScript. In addition, it modifies the command arguments specified in `MoorhenCarbohydrateValidation.tsx`, passing the `selectedMolecule.molNo` directly for validation rather than relying on the previous shim method.
